Feature #13271
Migrate Silverpeas Core Implementation Library to JakartaEE 10

+++ core-library/src/main/java/org/silverpeas/core/notification/user/delayed/DelayedNotificationManager.java
@@ -38,8 +38,8 @@ import org.silverpeas.core.util.CollectionUtil;
 import org.silverpeas.core.util.DateUtil;
 import org.silverpeas.core.util.MapUtil;
 
-import javax.inject.Inject;
-import javax.transaction.Transactional;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashSet;
@@ -150,7 +150,7 @@ public class DelayedNotificationManager implements DelayedNotification {
 
   @Override
   public int deleteDelayedNotifications(final Collection<Long> ids) {
-    int nbDeletes = 0;
+    long nbDeletes = 0;
     if (CollectionUtil.isNotEmpty(ids)) {
       for (final Collection<Long> idLot : CollectionUtil.split(ids)) {
         nbDeletes += dnRepository.deleteByIds(idLot);
@@ -158,7 +158,7 @@ public class DelayedNotificationManager implements DelayedNotification {
       nrRepository.deleteResources();
       dnRepository.flush();
     }
-    return nbDeletes;
+    return Math.toIntExact(nbDeletes);
   }
 
   /*
